Semi-automatic code cleanup, part 2

+++ community/kernel/src/main/java/org/neo4j/unsafe/impl/batchimport/cache/idmapping/string/LongCollisionValues.java
@@ -47,7 +47,7 @@ public class LongCollisionValues implements CollisionValues
     @Override
     public Object get( long offset )
     {
-        return (long) cache.get( offset );
+        return cache.get( offset );
     }
 
     @Override
